Remove unused files and reorganize code/config. Bring server url into .env
