+++ src/component/map/mapLayer.tsx
@@ -1,6 +1,7 @@
+import { env } from "@/service/env";
 import {
   DEFAULT_LATITUDE, DEFAULT_LONGITUDE, DEFAULT_MAP_SETTING, DEFAULT_VIEW_STATE, DEFAULT_ZOOM
-} from "@/util/constant";
+} from "@/util/config";
 import type { LatLng } from "@/util/general.types";
 import "mapbox-gl/dist/mapbox-gl.css";
 import React, { useState } from "react";
@@ -36,7 +37,7 @@ const MapLayer: React.FC<Props> = (props) => {
       {...viewState}
       {...DEFAULT_MAP_SETTING}
       mapStyle="mapbox://styles/mapbox/dark-v10"
-      mapboxAccessToken={import.meta.env.REACT_APP_MAPBOX_TOKEN}
+      mapboxAccessToken={env.REACT_APP_MAPBOX_TOKEN}
       onMove={event => setViewState(event.viewState)}
     >
       <FullscreenControl position="bottom-right" />
